[ELY-2506] Remove sshd-common dependency from wildfly-elytron-x500-cert module
