+++ ELY_Messages.txt
@@ -127,7 +127,7 @@
 29000 - 29999
 30000 - 30999
 # Reserved For master / Elytron 2
-40000 - 40999 master
+40000 - 40999 wildfly-elytron-ssh-util
 41000 - 41999 master
 42000 - 42999 master
 43000 - 43999 master
